+++ src/components/landing/jumbo/index.jumbo.tsx
@@ -3,17 +3,13 @@ import { useContext, useMemo, FC } from "react";
 import Image from "next/image";
 import Link from "next/link";
 import BouncingArrow from "@/components/landing/jumbo/bouncing-arrow";
-import {
-  JumboButtonKeys,
-  Roles,
-  jumboButtons,
-} from "@/components/landing/jumbo/constants";
 import ButtonGroup from "@/components/shared/button-group";
 import Scrolling from "@/components/shared/scrolling";
 import { ModalContext } from "@/context/modal/modal.context";
 import useMobileDetect from "@/hooks/use-mobile-detect";
 import { LinkIcon, StarIcon } from "@/icons";
 import styles from "@/styles/jumbo.style";
+import { JumboButtonKeys, Roles, jumboButtons } from "@/utils/constants";
 
 export type IRolesItems = {
   key: keyof typeof Roles;
