+++ src/utils/features.util.ts
@@ -1,6 +1,8 @@
 export enum Features {
   Follow = "follow",
+  Ministries = "ministries",
   Testimonials = "testimonials",
+  Donate = "donate",
 }
 
 export const features = {
@@ -9,9 +11,20 @@ export const features = {
     status: false,
     reason: "configure social media handles",
   },
+  [Features.Ministries]: {
+    title: Features.Ministries,
+    status: false,
+    reason: "fill in ministries with real values",
+  },
   [Features.Testimonials]: {
     title: Features.Testimonials,
     status: false,
-    reason: "fill in testimonials",
+    reason: "fill in testimonials with real values",
+  },
+  [Features.Donate]: {
+    title: Features.Donate,
+    status: false,
+    reason:
+      "configure the following production env vars: NEXT_PUBLIC_FLUTTERWAVE_DONATION_LINK, NEXT_PUBLIC_PAYSTACK_DONATION_LINK, and NEXT_PUBLIC_PAYSTACK_DONATION_LINK_REDIRECT",
   },
 };
